refactor: move useSearchbar into components/Searchbar folder
