+++ frontend/src/components/Searchbar/index.ts
@@ -1,1 +1,2 @@
 export { Searchbar as default } from './Searchbar'
+export { useSearchbar } from './useSearchbar'
