+++ frontend/src/components/Searchbar/useSearchbar.ts
